Create random category

+++ snippets/random/random.cpp
@@ -29,9 +29,9 @@ int main() {
     // Generating a histogram
     vector<int> hist(10, 0);
     for (int i = 0; i < 300; ++i) {
-        auto bin = static_cast<size_t>(d2(g2) + static_cast<double>(hist.size()) / 2);
-        if (bin >= 0 && bin < hist.size()) {
-            ++hist[bin];
+        auto bin = d2(g2) + static_cast<double>(hist.size()) / 2;
+        if (bin >= 0. && static_cast<size_t>(bin) < hist.size()) {
+            ++hist[static_cast<size_t>(bin)];
         }
     }
     for (int n : hist) {
